requirements.txt
fastapi
uvicorn
gradio_client
pymongo
python-dotenv
